+++ src/button/button.types.ts
@@ -1,0 +1,3 @@
+export type ButtonType = "primary" | "secondary" | "tertiary" | "ghost" | "danger" | "danger--primary" | "toolbar-action";
+
+export type ButtonSize = "sm" | "field";
