+++ src/button/index.ts
@@ -1,2 +1,3 @@
-export { Button, ButtonSize, ButtonType } from "./button.directive";
+export { Button } from "./button.directive";
 export { ButtonModule } from "./button.module";
+export { ButtonSize, ButtonType } from "./button.types";
