+++ src/checkbox/checkbox.types.ts
@@ -1,0 +1,1 @@
+export type CheckboxValue = boolean | "on" | "off";
