+++ src/checkbox/index.ts
@@ -1,2 +1,3 @@
-export * from "./checkbox.component";
-export * from "./checkbox.module";
+export { Checkbox, CheckboxState, CheckboxChange } from "./checkbox.component";
+export { CheckboxModule } from "./checkbox.module";
+export { CheckboxValue } from "./checkbox.types";
