+++ src/table/body/table-body.component.ts
@@ -7,7 +7,7 @@ import {
 import { TableModel } from "../table-model.class";
 import { I18n, Overridable } from "carbon-components-angular/i18n";
 import { Observable } from "rxjs";
-import { TableRowSize } from "../table.component";
+import { TableRowSize } from "../table.types";
 
 @Component({
 	// tslint:disable-next-line: component-selector
