+++ src/table/body/table-row.component.ts
@@ -10,7 +10,7 @@ import { TableModel } from "../table-model.class";
 import { I18n, Overridable } from "carbon-components-angular/i18n";
 import { TableItem } from "../table-item.class";
 import { Observable } from "rxjs";
-import { TableRowSize } from "../table.component";
+import { TableRowSize } from "../table.types";
 
 @Component({
 	// tslint:disable-next-line: component-selector
