+++ src/table/cell/table-checkbox.component.ts
@@ -8,7 +8,7 @@ import { I18n, Overridable } from "carbon-components-angular/i18n";
 import { TableItem } from "../table-item.class";
 import { TableRow } from "../table-row.class";
 import { Observable } from "rxjs";
-import { TableRowSize } from "../table.component";
+import { TableRowSize } from "../table.types";
 
 @Component({
 	// tslint:disable-next-line: component-selector
