+++ src/table/head/table-head-checkbox.component.ts
@@ -7,7 +7,7 @@ import {
 } from "@angular/core";
 import { I18n, Overridable } from "carbon-components-angular/i18n";
 import { Observable } from "rxjs";
-import { TableRowSize } from "../table.component";
+import { TableRowSize } from "../table.types";
 
 @Component({
 	// tslint:disable-next-line: component-selector
