+++ src/table/table.types.ts
@@ -1,0 +1,9 @@
+export interface TableTranslations {
+	FILTER: string;
+	END_OF_DATA: string;
+	SCROLL_TOP: string;
+	CHECKBOX_HEADER: string;
+	CHECKBOX_ROW: string;
+}
+
+export type TableRowSize = "sm" | "sh" | "md" | "lg";
