+++ src/utils/element.types.ts
@@ -1,0 +1,3 @@
+export interface ElementVisibilityEvent {
+	visible: boolean;
+}
